⚡  #51 FIX: 성분 기능 개선
Merge pull request #58 from E-GaeMuYac/#51/Materials

+++ src/models/ingredients.js
@@ -41,7 +41,7 @@ module.exports = (sequelize, DataTypes) => {
       },
       volume: {
         allowNull: false,
-        type: DataTypes.INTEGER,
+        type: DataTypes.DECIMAL(10, 5),
       },
       createdAt: {
         allowNull: false,
